# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'My API Documentation'
  description: 'API documentation for my Laravel application'
  contact:
    email: admin@example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'API Server'
paths:
...
  /api/wishlist:
    get:
      tags:
        - Wishlist
      summary: 'Get all wishlist items for the authenticated user'
      operationId: getWishlist
      responses:
        200:
          description: 'Wishlist items retrieved successfully — returns empty array if wishlist is empty'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        user_id:
                          type: integer
                          example: 1
                        product_id:
                          type: integer
                          example: 3
                        created_at:
                          type: string
                          format: date-time
                          example: '2024-01-01T00:00:00.000000Z'
                        updated_at:
                          type: string
                          format: date-time
                          example: '2024-01-01T00:00:00.000000Z'
                        product:
                          properties:
                            id:
                              type: integer
                              example: 3
                            name:
                              type: string
                              example: 'Rose Perfume'
                            price:
                              type: number
                              format: float
                              example: 29.99
                            image:
                              type: string
                              example: products/abc.jpg
                              nullable: true
                            stock:
                              type: integer
                              example: 100
                          type: object
                      type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
      security:
        -
          sanctum: []
    post:
      tags:
        - Wishlist
      summary: 'Add a product to the wishlist — ignored if product is already in wishlist'
      operationId: addToWishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
              properties:
                product_id:
                  description: 'Required — must exist in products table'
                  type: integer
                  example: 3
              type: object
      responses:
        201:
          description: 'Product added to wishlist successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    description: 'Newly created wishlist entry — note: product is NOT eager loaded on this response'
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      product_id:
                        type: integer
                        example: 3
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        409:
          description: "Conflict — product is already in the user's wishlist"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Already in wishlist'
                type: object
        422:
          description: 'Validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The product_id field is required.'
                  errors:
                    properties:
                      product_id:
                        type: array
                        items:
                          type: string
                          example: 'The selected product_id is invalid.'
                    type: object
                type: object
      security:
        -
          sanctum: []
  '/api/wishlist/{id}':
    delete:
      tags:
        - Wishlist
      summary: 'Remove a specific item from the wishlist'
      operationId: removeFromWishlist
      parameters:
        -
          name: id
          in: path
          description: 'Wishlist item ID — must belong to the authenticated user'
          required: true
          schema:
            type: integer
            example: 1
            minimum: 1
      responses:
        200:
          description: 'Item removed from wishlist successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Removed from wishlist'
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        404:
          description: 'Wishlist item not found or does not belong to the authenticated user'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Not found'
                type: object
      security:
        -
          sanctum: []
components:
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Enter token in format: Bearer <token>'
      bearerFormat: JWT
      scheme: bearer
...